optify/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
google-generativeai
python-dotenv